requirements.txt
pandas
cython==0.29.16
numpy
torch
torchvision
cython
pycocotools
tabulate
alfred-py
opencv-python
termcolor
cloudpickle
imagesize
portalocker
easydict
matplotlib
tensorboard==1.14
ngraph-core
ngraph-onnx
tqdm
pyzmq